c8y-operations/RemoteAccessConnect/requirements.txt
paho-mqtt
websocket_client
certifi
c8y-device-proxy